feat(Posting):Added Feed Logic
