+++ .env
@@ -1,0 +1,1 @@
+GOOGLE_APPLICATION_CREDENTIALS=C:\Users\bluev\Documents\itbsa-project\firebase.json
