--- api/logging/loggingApiInfo.ts
@@ -1,5 +1,0 @@
-import { ILoggingInfo } from './model/LoggingInfo';
-
-export const LOGGING_INFO: ILoggingInfo = {
-  contextPath: '/api',
-};
