--- api/logging/model/LoggingInfo.ts
@@ -1,3 +1,0 @@
-export interface ILoggingInfo {
-  contextPath: string;
-}
