+++ libs/db/FeedDB.ts
@@ -5,28 +5,18 @@ import type {
   WriteResult,
 } from 'firebase-admin/firestore';
 import type { DecodedIdToken } from 'firebase-admin/auth';
-import ILoggingDB from '../../interfaces/ILoggingDB';
+import IFeedDB from '../../interfaces/IFeedDB';
 import type { LoggingData } from '../../types/types';
-class LoggingDB implements ILoggingDB {
+class FeedDB implements IFeedDB {
   private db: Firestore;
 
   constructor() {
     this.db = admin.firestore();
   }
 
-  // private initializeFirebase(): App {
-  //   try {
-  //     return admin.initializeApp({
-  //       credential: admin.credential.cert(firebaseConfig as ServiceAccount),
-  //     });
-  //   } catch (e: any) {
-  //     throw new Error(e);
-  //   }
-  // }
-
   public async set(data: LoggingData): Promise<WriteResult> {
     try {
-      const collectionRef: CollectionReference = this.db.collection('logs');
+      const collectionRef: CollectionReference = this.db.collection('feed');
       return await collectionRef.doc().set(data);
     } catch (e: any) {
       throw new Error(e);
@@ -34,7 +24,7 @@ class LoggingDB implements ILoggingDB {
   }
   public async getByID(id: string): Promise<LoggingData | null> {
     try {
-      const collectionRef: CollectionReference = this.db.collection('logs');
+      const collectionRef: CollectionReference = this.db.collection('feed');
       const docRef = await collectionRef.doc(id).get();
       if (docRef.exists) {
         return docRef.data() as LoggingData;
@@ -49,7 +39,7 @@ class LoggingDB implements ILoggingDB {
     data: Partial<LoggingData>
   ): Promise<WriteResult> {
     try {
-      const collectionRef: CollectionReference = this.db.collection('logs');
+      const collectionRef: CollectionReference = this.db.collection('feed');
       return await collectionRef.doc(logReference).update(data);
     } catch (e: any) {
       throw new Error(e);
@@ -57,7 +47,7 @@ class LoggingDB implements ILoggingDB {
   }
   public async deleteByID(id: string): Promise<Boolean> {
     try {
-      const collectionRef: CollectionReference = this.db.collection('logs');
+      const collectionRef: CollectionReference = this.db.collection('feed');
       return !!(await collectionRef.doc(id).delete());
     } catch (e: any) {
       throw new Error(e);
@@ -73,4 +63,4 @@ class LoggingDB implements ILoggingDB {
   }
 }
 
-export default LoggingDB;
+export default FeedDB;
