+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "firebase-admin-microservice",
+  "name": "feed-service",
   "version": "1.0.0",
   "main": "app.ts",
   "repository": "git@github.com:Milahn-Henri-Louis-Irwin-ITBSA/firebase-admin-microservice.git",
@@ -7,16 +7,17 @@
   "license": "MIT",
   "scripts": {
     "dev": "tsc & nodemon api/app.ts",
-    "prod": "tsc"
+    "prod": "rm -r dist && tsc"
   },
   "dependencies": {
     "axios": "^1.4.0",
     "body-parser": "^1.20.2",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
-    "express-rate-limit": "^7.1.1",
-    "firebase-admin": "^11.11.0",
+    "express-rate-limit": "^7.1.3",
+    "firebase-admin": "^11.10.1",
     "module-alias": "^2.2.3",
     "reflect-metadata": "^0.1.13",
     "routing-controllers": "^0.10.4",
@@ -25,7 +26,11 @@
     "typedi": "^0.10.0"
   },
   "devDependencies": {
+    "@types/chai": "^4.3.5",
+    "@types/mocha": "^10.0.1",
     "@types/node": "^20.4.2",
+    "chai": "^4.3.8",
+    "mocha": "^10.2.0",
     "nodemon": "^3.0.1",
     "typescript": "^5.1.6"
   }
