+++ types/types.d.ts
@@ -1,9 +1,7 @@
 import type { Timestamp } from 'firebase-admin/firestore';
-import type { UserInfo } from 'firebase-admin/auth';
+import type { DecodedIdToken } from 'firebase-admin/auth';
 export type LoggingData = {
   created_at: Timestamp;
-  type: string;
-  written_by: string;
   message: string;
-  user_uid: UserInfo;
+  created_by: DecodedIdToken['uid'];
 };
